Add and remove a div that encompasses all elements
Firstly add div that contains all the content, then remove it as it
wasn't needed anymore because display of flex and other declarations
that help to position three main containers were added to body element.
In the end, index.html file is the same, just some new spaces here and
there.